Add new Finder feature
This change adds a Finder factory, which allows creating handlers
for command arguments. The factory takes in a scope and a target
trait, and creates a finder that, when parameterized with the actor
and a name string, searches for a matching target from the
perspective of the actor.

Example:

```
const containersInRoom = Finder.create({ scope: "room", target: "container" })
```