Add basic yawn command
This command is added as exploration. As it's the simplest form of
command; an emote taking no arguments, and having no side effects.

The test revealed that there needs to be some clarification around
sensory events, and how they resolve. Questions like: "should a
room know how to emit events, or should events know how to emit in
rooms", still need to be answered. Events also need to be able to
tell when they are resolved in the actor where they originated, and
adjust their message accordingly.